Move simplifyRollFormula into seperate file, fix error with highlightCriticalFailure if no dice in formula
