+++ module/applications/item/ability-use-dialog.mjs
@@ -65,7 +65,7 @@ export default class AbilityUseDialog extends Dialog {
             label: label,
             callback: html => {
               const fd = new FormDataExtended(html[0].querySelector("form"));
-              resolve(fd.toObject());
+              resolve(fd.object);
             }
           }
         },
