+++ module/dice/_module.mjs
@@ -1,3 +1,4 @@
 export {default as D20Roll} from "./d20-roll.mjs";
 export {default as DamageRoll} from "./damage-roll.mjs";
 export * from "./dice.mjs";
+export {default as simplifyRollFormula} from "./simplify-roll-formula.mjs";
